Docstrings

GuiggianiRichardsonDuffy.ANALYTICAL_KERNELS — Constant

const ANALYTICAL_KERNELS = [:LaplaceHypersingular, :ElastostaticHypersingular]

Available kernels with analytical Laurent coefficients.

GuiggianiRichardsonDuffy.EXPANSION_METHODS — Constant

const EXPANSION_METHODS = [:analytical, :auto_diff, :semi_richardson, :full_richardson]

Available expansion methods for Laurent coefficients of singular kernels.

GuiggianiRichardsonDuffy.ElastostaticHypersingularClosedFormF₋₁ — Method

ElastostaticHypersingularClosedFormF₋₁(args...; kwargs...)

Compute only F₋₁ the old way (analytically with F₋₂) such that F(ρ, θ) := ρ × J × Nᵖ × Vᵢₖⱼ = F₋₁ / ρ + F₋₂ / ρ² + O(1), for the elastostatic hypersingular kernel.

GuiggianiRichardsonDuffy.ElastostaticHypersingularClosedFormF₋₂ — Method

ElastostaticHypersingularClosedFormF₋₂(args...; kwargs...)

Compute only F₋₂ the old way (analytically with F₋₁) such that F(ρ, θ) := ρ × J × Nᵖ × Vᵢₖⱼ = F₋₁ / ρ + F₋₂ / ρ² + O(1), for the elastostatic hypersingular kernel.

GuiggianiRichardsonDuffy.LaplaceHypersingularClosedFormF₋₁ — Method

LaplaceHypersingularClosedFormF₋₁(args...; kwargs...)

Compute only F₋₁ the old way (analytically with F₋₂) such that F(ρ, θ) := ρ × J × Nᵖ × Vᵢ = F₋₁ / ρ + F₋₂ / ρ² + O(1), for Laplace hypersingular kernel.

GuiggianiRichardsonDuffy.LaplaceHypersingularClosedFormF₋₂ — Method

LaplaceHypersingularClosedFormF₋₂(args...; kwargs...)

Compute only F₋₂ the old way (analytically with F₋₁) such that F(ρ, θ) := ρ × J × Nᵖ × Vᵢ = F₋₁ / ρ + F₋₂ / ρ² + O(1), for Laplace hypersingular kernel.

GuiggianiRichardsonDuffy.guiggiani_singular_integral — Method

guiggiani_singular_integral(
	K,
	û,
	x̂,
	el::Inti.ReferenceInterpolant,
	n_rho,
	n_theta;
	sorder::Val{P} = Val(-2),
	expansion::Symbol = :full_richardson,
	kernel_kwargs::NamedTuple = NamedTuple(),
	richardson_kwargs::NamedTuple = NamedTuple(),
	kwargs...,
) where {P}

Given a kernel K, a function û defined on the reference element el, a point x̂ on the reference element where the singularity is located, the number of quadrature points in the radial direction n_rho, the number of quadrature points in the angular direction n_theta, and the order of the singularity sorder (which has to be -1, -2 or -3), computes the integral of the kernel over the reference element using the Guiggiani method using expansion of Laurent coefficients specified in the expansion argument.

See GuiggianiRichardsonDuffy.laurents_coeffs for the available expansion methods and their parameters.

GuiggianiRichardsonDuffy.laurents_coeffs — Method

laurents_coeffs(K, el::Inti.ReferenceInterpolant, û, x̂; expansion = (method = :full_richardson,), kwargs...)

Given a kernel K, a reference element el, a function û defined on the reference element, and a point x̂ on the reference element, returns the laurent coefficients F₋₂ and F₋₁ for the kernel K in polar coordinates centered at x̂. The coefficients are computed using the method specified in the expansion argument, which can be one of the following:

:analytical: uses analytical expressions for the coefficients (if available). kernel_kwargs... are passed to analytical functions.

:auto_diff: uses semi-analytical expressions for the coefficients (if available i.e. when the property of the kernel being translation-invariant holds) based on automatic differentiation used in the ForwardDiff.jl package. kernel_kwargs... are passed to the kernel K̂.

:semi_richardson: uses another semi-analytical method for the coefficients (if available i.e. when the property of the kernel being translation-invariant holds). richardson_kwargs... are passed to Richardson.extrapolate (see Richardson.jl) and kernel_kwargs... are passed to the kernel K̂.

:full_richardson: uses Richardson extrapolation to compute both coefficients, available by default for any kernel. richardson_kwargs... are passed to Richardson.extrapolate (see Richardson.jl) and kernel_kwargs... are passed to the kernel K̂.

K has to be called as K(qx, qy, r̂; kernelkwargs...) where r̂ is the normalized relative position vector, qx = (coords = x, normal = nx) and qy = (coords = y, normal = ny). K(qx, qy, r̂; kernelkwargs...) is returning the tuple (1/rˢ, K̂(qx, qy, r̂; kernel_kwargs...)) where s is the order of the singularity.

You can also put all the keyword arguments in kwargs..., they will be automatically split between kernel and richardson extrapolation arguments, which are in general : first_contract, contract, breaktol, maxeval, atol, rtol, x0, described in the Richardson.extrapolate documentation (see Richardson.jl).

GuiggianiRichardsonDuffy.polar_kernel_fun — Method

polar_kernel_fun(K::Inti.AbstractKernel, el::Inti.ReferenceInterpolant, û, x̂)

Given a kernel K, a reference element el, a function û defined on the reference element, and a point x̂ on the reference element, returns a function F that computes the complete kernel in polar coordinates centered at x̂ : F(ρ, θ) = K(x, y) * J(ŷ) * ρ * û(ŷ) where x = el(x̂), ŷ = x̂ + ρ * (cos(θ), sin(θ)), y = el(ŷ), and J(ŷ) is the integration measure at ŷ. F will be called as F(ρ, θ). K has to be called as K(qx, qy) where qx = (coords = x, normal = nx) and qy = (coords = y, normal = ny) are cartesian points with their normals.

K must return a unique value.

GuiggianiRichardsonDuffy.rho_fun — Method

rho_fun(ref_domain::Inti.ReferenceDomain, x̂)

Given a reference domain ref_domain and a point x̂ in the reference domain, returns the function ρ(θ) that gives the distance from x̂ to the boundary of the reference domain in the direction θ. ρ will be called as ρ(θ).